requirements.txt
PyYAML==6.0
svgwrite==1.4.3